+++ C2_Profiles/github/github/c2_code/config.py
@@ -1,0 +1,1 @@
+config = {}
