+++ C2_Profiles/github/requirements.txt
@@ -1,5 +1,3 @@
 aiohttp
-flask
 mythic_container
 quart
-requests
